Added zipper for building project

+++ server/build.js
@@ -1,15 +1,38 @@
-const {createDir, copyFile, copyDir, removeDir} = require('./utils/directory');
-const {compileSass} = require('./utils/compiler');
+const {
+  createDir,
+  copyFile,
+  copyDir,
+  removeDir,
+} = require('./utils/directory');
+const { zipper } = require('./utils/zipper');
+const { compileSass } = require('./utils/sass');
+const config = require('../config.json');
 
-const DIRECTORY = './build'
+const DOMAIN = [
+  'Plan',
+  'Recipe',
+  'GeneraBlog',
+  'Ingredient',
+  'Nutrient',
+  'Wiki',
+];
+
+if (!config.title || !config.domain)
+  throw Error('You must declare title and domain in the config.json file');
+if (!DOMAIN.includes(config.domain))
+  throw Error(
+    `Incorrect domain: You must choose options between ${DOMAIN.join(', ')}`,
+  );
+
+const DIRECTORY = './build';
 
 // Generating the build folder
 
 // 1. Removing the existing build folder
-removeDir(DIRECTORY)
+removeDir(DIRECTORY);
 
 // 2. Creating a new build folder
-createDir(DIRECTORY)
+createDir(DIRECTORY);
 
 // 3. Copying all the required files into build folder
 copyFile('./config.json', DIRECTORY);
@@ -17,12 +40,6 @@ copyFile('./index.ejs', DIRECTORY);
 
 copyDir('./assets', DIRECTORY);
 
+compileSass(DIRECTORY);
 
-compileSass(DIRECTORY)
-
-console.log("Build successfully")
-
-
-
-
-
+zipper(DIRECTORY);
